+++ nibabel/testing/helpers.py
