+++ nibabel/tests/test_image_api.py
@@ -45,8 +45,8 @@ import unittest
 import pytest
 
 from numpy.testing import assert_almost_equal, assert_array_equal, assert_warns, assert_allclose
-from ..testing_pytest import (bytesio_round_trip, bytesio_filemap,
-                              assert_data_similar, clear_and_catch_warnings)
+from ..testing import (bytesio_round_trip, bytesio_filemap,
+                       assert_data_similar, clear_and_catch_warnings)
 from ..tmpdirs import InTemporaryDirectory
 from ..deprecator import ExpiredDeprecationError
